update names to better handle multiple examples e.g. "local" duplicate
